+++ src/components/home/ImportChallengeCard.tsx
@@ -3,51 +3,10 @@ import ImportImage from "../../assets/import.png"
 import Button from '@mui/material/Button';
 import { Ember } from "../../emberCommunication";
 import { useNavigate } from "react-router-dom";
-import simpleTypeGuard, { SimpleArray, SimpleStringOptional, SimpleBoolean, SimpleNumber, SimpleString, SimpleBooleanOptional, SimpleObjectOptional, SimpleExactMatch } from 'simple-type-guard';
 import { useState } from "react";
 import { Modal, Paper, Typography } from "@mui/material";
 import { useTranslation } from "react-i18next";
-
-
-type SceneType = "lita" | "duba" | "toto" | "coty" | "niandu" | "pinguine" | "yaguarete" | "carpincho" | "custom"
-const sceneTypes: SceneType[] =  ["lita", "duba", "toto", "coty", "niandu", "pinguine", "yaguarete", "carpincho", "custom"] //Used for file validity checking
-
-type Cell = "P" | "-" | "A" | "" // should consider different objects / empty cell / no cell at all
-const cells: Cell[] = ["P", "-", "A", ""] //Used for file validity checking
-
-type SceneMap = Cell[]
-
-export type Scene = {
-    type: SceneType
-    maps: SceneMap[]
-}
-
-type Assesments =  {
-    itWorks?: boolean, // old "debeFelicitar", default true
-    decomposition?: DecompositionAssessment,
-    simpleRepetition?: boolean,
-    conditionalRepetition?: boolean,
-    conditionalAlternative?: boolean,
-}
-
-type DecompositionAssessment = { maxProgramLength: number }
-
-export type SerializedChallenge = {
-    fileVersion: number,
-    title: string,
-    statement: {
-        description: string,
-        clue?: string
-    },
-    scene: Scene,
-    toolbox: {
-        blocks: string[], // for now, block ids, future: could be objects.
-        categorized?: boolean // default true
-    },
-    stepByStep?: boolean, // default false
-    predefinedSolution?: string,
-    assesments?: Assesments
-}
+import { SerializedChallenge, isValidChallenge } from "../serializedChallenge";
 
 export const ImportChallengeCard = () => {
     const navigate = useNavigate();
@@ -62,30 +21,6 @@ export const ImportChallengeCard = () => {
         setModalOpen(true)
     }
 
-    const isValidChallenge = (json: unknown): boolean => 
-        simpleTypeGuard<SerializedChallenge>(json, {
-            fileVersion: SimpleNumber,
-            title: SimpleString,
-            statement: {description: SimpleString, clue: SimpleStringOptional},
-            toolbox: {blocks: new SimpleArray(SimpleString), categorized: SimpleBooleanOptional},
-            stepByStep: SimpleBooleanOptional,
-            predefinedSolution: SimpleStringOptional,
-            scene: { 
-                type: SimpleString, //TODO: Poner los exact match
-                maps: new SimpleArray<Cell[]>(new SimpleArray<Cell>(SimpleString)) //TODO: Poner los exact match
-            },
-            assesments: new SimpleObjectOptional<Assesments>({
-                itWorks: SimpleBooleanOptional,
-                simpleRepetition: SimpleBooleanOptional,
-                conditionalRepetition: SimpleBooleanOptional,
-                conditionalAlternative: SimpleBooleanOptional,
-                decomposition: new SimpleObjectOptional<DecompositionAssessment>({
-                    maxProgramLength: SimpleNumber
-                })
-
-            })
-        })
-
     const readFile = async (event: any) => {
         const file: File = event.target.files[0]
         const content: string = await file.text()
@@ -94,7 +29,7 @@ export const ImportChallengeCard = () => {
         event.target.value = null // Without this Chrome seems to cache the file and prevents reruns of this function. 
 
         if (isValidChallenge(challengeJson)) {
-            goToChallenge(challengeJson as SerializedChallenge)
+            goToChallenge(challengeJson)
         }
         else {
             showErrorModal()
